CB-2927 read ssh key in initConnection
