+++ server/bundles/io.cloudbeaver.server/src/io/cloudbeaver/service/core/impl/WebServiceCore.java
@@ -337,9 +337,9 @@ public class WebServiceCore implements DBWServiceCore {
                 if (CommonUtils.toBoolean(c.isSavePassword()) && !CommonUtils.isEmpty(c.getUserName())) {
                     DBWHandlerConfiguration handlerCfg = dataSourceContainer.getConnectionConfiguration().getHandler(c.getId());
                     if (handlerCfg != null) {
-                        handlerCfg.setUserName(c.getUserName());
-                        handlerCfg.setPassword(c.getPassword());
+                        WebDataSourceUtils.updateHandlerConfig(handlerCfg, c);
                         handlerCfg.setSavePassword(true);
+                        handlerCfg.setEnabled(true);
                         saveConfig[0] = true;
                     }
                 }
@@ -650,7 +650,7 @@ public class WebServiceCore implements DBWServiceCore {
                     monitor.subTask("Initialize tunnel");
 
                     DBWHandlerConfiguration configuration = new DBWHandlerConfiguration(handlerDescriptor, null);
-                    WebServiceUtils.updateHandlerConfig(configuration, nhConfig);
+                    WebDataSourceUtils.updateHandlerConfig(configuration, nhConfig);
                     configuration.setSavePassword(true);
                     configuration.setEnabled(true);
                     tunnel.initializeHandler(monitor, configuration, connectionConfig);
